LayoutTests/http/tests/site-isolation/sandbox-allow-top-navigation-cross-origin-iframe.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script src="/js-test-resources/js-test.js"></script>
<script>

description("Verifies a sandboxed cross-origin iframe with allow-top-navigation can navigate its parent popup to a blob URL.");
jsTestIsAsync = true;

function runTest() {
    let popup = window.open();
    let iframe = popup.document.createElement('iframe');
    iframe.sandbox = "allow-scripts allow-same-origin allow-top-navigation";
    iframe.src = "http://localhost:8000/site-isolation/resources/navigate-parent-to-blob.html";

    window.addEventListener("message", (event) => {
        if (event.source !== popup)
            return;
        shouldBeEqualToString("event.data", "blob loaded");
        popup.close();
        finishJSTest();
    });

    popup.document.body.appendChild(iframe);
}

function runTestIfInTestRunner() { if (window.testRunner) { runTest() } }
</script>
<body onload="runTestIfInTestRunner()">
<button onclick="runTest()">click to run test manually in a browser</button>
</body>
